Split autodoc signature utilities to a new module
