+++ tests/test_ext_autodoc/test_ext_autodoc.py
@@ -25,18 +25,10 @@ from sphinx.ext.autodoc._directive_options import (
 from sphinx.ext.autodoc._docstrings import _get_docstring_lines
 from sphinx.ext.autodoc._documenters import Documenter
 from sphinx.ext.autodoc._generate import _generate_directives
-from sphinx.ext.autodoc._property_types import (
-    _ClassDefProperties,
-    _FunctionDefProperties,
-    _ItemProperties,
-)
+from sphinx.ext.autodoc._property_types import _ItemProperties
 from sphinx.ext.autodoc._sentinels import ALL
 from sphinx.ext.autodoc.directive import _AutodocAttrGetter
-from sphinx.ext.autodoc.importer import (
-    _format_signatures,
-    _load_object_by_name,
-    _parse_name,
-)
+from sphinx.ext.autodoc.importer import _load_object_by_name, _parse_name
 from sphinx.util.inspect import safe_getattr
 
 from tests.test_ext_autodoc.autodoc_util import do_autodoc
@@ -112,309 +104,6 @@ def test_parse_name(caplog: pytest.LogCaptureFixture) -> None:
     assert parsed == ('sphinx.testing.util', ['SphinxTestApp', 'cleanup'], None, None)
     parsed = parse('method', 'SphinxTestApp.cleanup')
     assert parsed == ('sphinx.testing.util', ['SphinxTestApp', 'cleanup'], None, None)
-
-
-def format_sig(obj_type, name, obj, *, app, args=None, retann=None):
-    get_attr = _AutodocAttrGetter(app.registry.autodoc_attrgetters)
-    options = _AutoDocumenterOptions()
-
-    parent = object  # dummy
-    props = _ClassDefProperties(
-        obj_type=obj_type,
-        module_name='',
-        parts=(name,),
-        docstring_lines=(),
-        bases=getattr(obj, '__bases__', None),
-        _obj=obj,
-        _obj___module__=getattr(obj, '__module__', None),
-        _obj___qualname__=getattr(obj, '__qualname__', None),
-        _obj___name__=name,
-        _obj_bases=(),
-        _obj_is_new_type=False,
-        _obj_is_typevar=False,
-    )
-    docstrings = _get_docstring_lines(
-        props,
-        class_doc_from=app.config.autoclass_content,
-        get_attr=get_attr,
-        inherit_docstrings=app.config.autodoc_inherit_docstrings,
-        parent=parent,
-        tab_width=8,
-    )
-    signatures = _format_signatures(
-        autodoc_annotations={},
-        config=app.config,
-        docstrings=docstrings,
-        events=app.events,
-        get_attr=get_attr,
-        options=options,
-        parent=parent,
-        props=props,
-        args=args,
-        retann=retann,
-    )
-    if not signatures:
-        return ()
-    assert len(signatures) == 1
-    return signatures[0]
-
-
-@pytest.mark.sphinx('html', testroot='root')
-def test_format_signatures(app):
-    def process_signature(_app, what, name, _obj, _options, _args, _retann):
-        processed_signatures.append((what, name))
-        if name == '.bar':
-            return '42', None
-        return None
-
-    def skip_member(_app, _what, name, _obj, skip, _options):
-        if name in {'__special1__', '__special2__'}:
-            return skip
-        if name.startswith('__'):
-            return True
-        if name == ('skipmeth', ''):
-            return True
-        return None
-
-    app.connect('autodoc-process-signature', process_signature)
-    app.connect('autodoc-skip-member', skip_member)
-
-    # no signatures for modules
-    assert format_sig('module', 'test', None, app=app) == ()
-
-    # test for functions
-    def f(a, b, c=1, **d):
-        pass
-
-    def g(a='\n'):
-        pass
-
-    assert format_sig('function', 'f', f, app=app) == ('(a, b, c=1, **d)', '')
-    assert format_sig('function', 'f', f, app=app, args='(a, b, c, d)') == (
-        '(a, b, c, d)',
-        '',
-    )
-    assert format_sig('function', 'g', g, app=app) == (r"(a='\n')", '')
-
-    if sys.version_info >= (3, 12):
-        for params, expect in [
-            ('(a=1)', '(a=1)'),
-            ('(a: int=1)', '(a: int = 1)'),  # auto whitespace formatting
-            ('(a:list[T]   =[], b=None)', '(a: list[T] = [], b=None)'),  # idem
-        ]:
-            ns = {}
-            exec(f'def f[T]{params}: pass', ns)  # NoQA: S102
-            f = ns['f']
-            assert format_sig('function', 'f', f, app=app) == (expect, '')
-            assert format_sig('function', 'f', f, app=app, args='(...)') == (
-                '(...)',
-                '',
-            )
-            assert format_sig(
-                'function', 'f', f, app=app, args='(...)', retann='...'
-            ) == ('(...)', '...')
-
-            exec(f'def f[T]{params} -> list[T]: return []', ns)  # NoQA: S102
-            f = ns['f']
-            assert format_sig('function', 'f', f, app=app) == (expect, 'list[T]')
-            assert format_sig('function', 'f', f, app=app, args='(...)') == (
-                '(...)',
-                '',
-            )
-            assert format_sig(
-                'function', 'f', f, app=app, args='(...)', retann='...'
-            ) == ('(...)', '...')
-
-    # TODO(picnixz): add more test cases for PEP-695 classes as well (though
-    # complex cases are less likely to appear and are painful to test).
-
-    # test for classes
-    class D:
-        pass
-
-    class E:
-        def __init__(self):
-            pass
-
-    # an empty init and no init are the same
-    for C in (D, E):
-        assert format_sig('class', 'D', C, app=app) == ('()', '')
-
-    class SomeMeta(type):
-        def __call__(cls, a, b=None):
-            return type.__call__(cls, a, b)
-
-    # these three are all equivalent
-    class F:
-        def __init__(self, a, b=None):
-            pass
-
-    class FNew:
-        def __new__(cls, a, b=None):  # NoQA: ARG004
-            return super().__new__(cls)
-
-    class FMeta(metaclass=SomeMeta):
-        pass
-
-    # and subclasses should always inherit
-    class G(F):
-        pass
-
-    class GNew(FNew):
-        pass
-
-    class GMeta(FMeta):
-        pass
-
-    # subclasses inherit
-    for C in (F, FNew, FMeta, G, GNew, GMeta):
-        assert format_sig('class', 'C', C, app=app) == ('(a, b=None)', '')
-    assert format_sig('class', 'C', D, app=app, args='(a, b)', retann='X') == (
-        '(a, b)',
-        'X',
-    )
-
-    class ListSubclass(list):  # NoQA: FURB189
-        pass
-
-    # only supported if the python implementation decides to document it
-    if getattr(list, '__text_signature__', None) is not None:
-        assert format_sig('class', 'C', ListSubclass, app=app) == (
-            '(iterable=(), /)',
-            '',
-        )
-    else:
-        assert format_sig('class', 'C', ListSubclass, app=app) == ()
-
-    class ExceptionSubclass(Exception):
-        pass
-
-    # Exception has no __text_signature__ at least in Python 3.11
-    if getattr(Exception, '__text_signature__', None) is None:
-        assert format_sig('class', 'C', ExceptionSubclass, app=app) == ()
-
-    # __init__ have signature at first line of docstring
-    app.config.autoclass_content = 'both'
-
-    class F2:
-        """some docstring for F2."""
-
-        def __init__(self, *args, **kw):
-            """
-            __init__(a1, a2, kw1=True, kw2=False)
-
-            some docstring for __init__.
-            """  # NoQA: D212
-
-    class G2(F2):
-        pass
-
-    assert format_sig('class', 'F2', F2, app=app) == (
-        '(a1, a2, kw1=True, kw2=False)',
-        '',
-    )
-    assert format_sig('class', 'G2', G2, app=app) == (
-        '(a1, a2, kw1=True, kw2=False)',
-        '',
-    )
-
-    # test for methods
-    class H:
-        def foo1(self, b, *c):
-            pass
-
-        def foo2(b, *c):  # NoQA: N805
-            pass
-
-        def foo3(self, d='\n'):
-            pass
-
-    assert format_sig('method', 'H.foo', H.foo1, app=app) == ('(b, *c)', '')
-    assert format_sig('method', 'H.foo', H.foo1, app=app, args='(a)') == ('(a)', '')
-    assert format_sig('method', 'H.foo', H.foo2, app=app) == ('(*c)', '')
-    assert format_sig('method', 'H.foo', H.foo3, app=app) == (r"(d='\n')", '')
-
-    # test bound methods interpreted as functions
-    assert format_sig('function', 'foo', H().foo1, app=app) == ('(b, *c)', '')
-    assert format_sig('function', 'foo', H().foo2, app=app) == ('(*c)', '')
-    assert format_sig('function', 'foo', H().foo3, app=app) == (r"(d='\n')", '')
-
-    # test exception handling (exception is caught and args is '')
-    app.config.autodoc_docstring_signature = False
-    assert format_sig('function', 'int', int, app=app) == ()
-
-    # test processing by event handler
-    assert format_sig('method', 'bar', H.foo1, app=app) == ('42', '')
-
-    # test functions created via functools.partial
-    from functools import partial
-
-    curried1 = partial(lambda a, b, c: None, 'A')
-    assert format_sig('function', 'curried1', curried1, app=app) == ('(b, c)', '')
-    curried2 = partial(lambda a, b, c=42: None, 'A')
-    assert format_sig('function', 'curried2', curried2, app=app) == ('(b, c=42)', '')
-    curried3 = partial(lambda a, b, *c: None, 'A')
-    assert format_sig('function', 'curried3', curried3, app=app) == ('(b, *c)', '')
-    curried4 = partial(lambda a, b, c=42, *d, **e: None, 'A')
-    assert format_sig('function', 'curried4', curried4, app=app) == (
-        '(b, c=42, *d, **e)',
-        '',
-    )
-
-
-@pytest.mark.sphinx('html', testroot='ext-autodoc')
-def test_autodoc_process_signature_typing_generic(app):
-    actual = do_autodoc(app, 'class', 'target.generic_class.A', {})
-
-    assert list(actual) == [
-        '',
-        '.. py:class:: A(a, b=None)',
-        '   :module: target.generic_class',
-        '',
-        '   docstring for A',
-        '',
-    ]
-
-
-@pytest.mark.sphinx('html', testroot='root')
-def test_autodoc_process_signature_typehints(app):
-    captured = []
-
-    def process_signature(*args):
-        captured.append(args)
-
-    app.connect('autodoc-process-signature', process_signature)
-
-    def func(x: int, y: int) -> int:  # type: ignore[empty-body]
-        pass
-
-    props = _FunctionDefProperties(
-        obj_type='function',
-        module_name='',
-        parts=('func',),
-        docstring_lines=(),
-        _obj=func,
-        _obj___module__=None,
-        _obj___qualname__=None,
-        _obj___name__=None,
-        properties=frozenset(),
-    )
-
-    get_attr = _AutodocAttrGetter(app.registry.autodoc_attrgetters)
-    options = _AutoDocumenterOptions()
-    _format_signatures(
-        autodoc_annotations={},
-        config=app.config,
-        docstrings=None,
-        events=app.events,
-        get_attr=get_attr,
-        options=options,
-        parent=None,
-        props=props,
-    )
-    assert captured == [
-        (app, 'function', '.func', func, options, '(x: int, y: int)', 'int')
-    ]
 
 
 def get_docstring_lines(obj_type, obj, *, config):
